Fix pod creation save
